Rename embedded form collection to container, and expect a form label
