--- lib/formalist/rich_text/embedded_form_collection.rb
@@ -1,9 +1,0 @@
-require "formalist/rich_text/embedded_form_collection/mixin"
-
-module Formalist
-  module RichText
-    class EmbeddedFormCollection
-      include Mixin
-    end
-  end
-end
