+++ lib/formalist/rich_text/embedded_forms_container.rb
@@ -1,0 +1,9 @@
+require "formalist/rich_text/embedded_forms_container/mixin"
+
+module Formalist
+  module RichText
+    class EmbeddedFormsContainer
+      include Mixin
+    end
+  end
+end
